Dev into main

See merge request I488027/s3-roomies-backend!20

+++ sh-backend/src/main/java/nl/fontys/s3/studenthousing/controller/dto/ListingDTO.java
@@ -14,7 +14,7 @@ public class ListingDTO {
     private String city;
     private String neighborhood;
     private String description;
-    private Integer surfaceArea; // In square meters
+    private Double surfaceArea; // In square meters
     private Double rent; // In Euro's
     private Boolean petsAllowed;
     private Boolean isActive;
