+++ sh-backend/src/main/resources/db/migration/V1__CreateTables.sql
@@ -23,7 +23,7 @@ CREATE TABLE listing(
     neighborhood varchar(100) NOT NULL,
     owner_id int,
     description varchar(10000) NOT NULL,
-    surface_area int NOT NULL,
+    surface_area double NOT NULL,
     rent double NOT NULL,
     pets_allowed boolean NOT NULL,
     end_date DATETIME NOT NULL,
